[x86] x86_64: Removed compilation warnings.

+++ arch/x86/cpu/x86_64/include/arch_types.h
@@ -25,10 +25,10 @@
 
 /** cpu specific types */
 typedef unsigned int irq_flags_t;
-typedef unsigned int virtual_addr_t;
-typedef unsigned int virtual_size_t;
-typedef unsigned int physical_addr_t;
-typedef unsigned int physical_size_t;
+typedef unsigned long virtual_addr_t; /* 64-bit virtual addresses */
+typedef unsigned long virtual_size_t;
+typedef unsigned long physical_addr_t;
+typedef unsigned long physical_size_t;
 typedef unsigned int clock_freq_t;
 typedef unsigned long long jiffies_t;
